separating out the plugin control plots
previously they were all in the ToyTree class, now
they have their own class ControlPlots.
