+++ core/src/MethodPluginScan.cpp
@@ -361,7 +361,10 @@ double MethodPluginScan::getPvalue1d(RooSlimFitResult* plhScan, double chi2minGl
 	computePvalue1d(plhScan, chi2minGlobal, myTree, id, myFit, myPb);
 
 	// compute p-value
-	if ( arg->controlplot ) myTree->ctrlPlotSummary();
+	if ( arg->controlplot ) {
+		ControlPlots cp(myTree);
+		cp.ctrlPlotSummary();
+	}
 	TH1F *h = analyseToys(myTree, id);
 	float scanpoint = plhScan->getParVal(scanVar1);
 	double pvalue = h->GetBinContent(h->FindBin(scanpoint));
@@ -942,13 +945,14 @@ void MethodPluginScan::readScan1dTrees(int runMin, int runMax)
 	t.open();
 
 	if ( arg->controlplot ) {
-		if ( arg->plotid==0 || arg->plotid==1 ) t.ctrlPlotMore(profileLH);
-		if ( arg->plotid==0 || arg->plotid==2 ) t.ctrlPlotSummary();
-		if ( arg->plotid==0 || arg->plotid==3 ) t.ctrlPlotNuisances();
-		if ( arg->plotid==0 || arg->plotid==4 ) t.ctrlPlotObservables();
-		if ( arg->plotid==0 || arg->plotid==5 ) t.ctrlPlotChi2Distribution();
-		if ( arg->plotid==0 || arg->plotid==6 ) t.ctrlPlotChi2Parabola();
-		t.saveCtrlPlots();
+		ControlPlots cp(&t);
+		if ( arg->plotid==0 || arg->plotid==1 ) cp.ctrlPlotMore(profileLH);
+		if ( arg->plotid==0 || arg->plotid==2 ) cp.ctrlPlotSummary();
+		if ( arg->plotid==0 || arg->plotid==3 ) cp.ctrlPlotNuisances();
+		if ( arg->plotid==0 || arg->plotid==4 ) cp.ctrlPlotObservables();
+		if ( arg->plotid==0 || arg->plotid==5 ) cp.ctrlPlotChi2Distribution();
+		if ( arg->plotid==0 || arg->plotid==6 ) cp.ctrlPlotChi2Parabola();
+		cp.saveCtrlPlots();
 	}
 
 	if ( hCL ) delete hCL;
